# backend/openapi.yaml
openapi: 3.0.0
info:
  title: GamePlatform API
  version: v1.0.0
  description: API for managing a gaming platform including championships, events, user registrations, results, and more.
servers:
  - url: /api # Assuming all backend routes are prefixed with /api
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
          description: A human-readable error message.
        details:
          type: string
          description: Optional additional details about the error.
        code:
          type: string
          description: Optional error code.
...
    UserProfile:
      type: object
      properties:
        id:
          type: string
          format: uuid
          description: User ID (from Supabase Auth).
        username:
          type: string
          nullable: true
        avatar_url:
          type: string
          format: url
          nullable: true
        website:
          type: string
          format: url
          nullable: true
        role:
          type: string
          enum: [user, admin, moderator] # Example roles
          default: user
        updated_at:
          type: string
          format: date-time
    UserUpdate:
      type: object
      properties:
        username:
          type: string
        avatar_url:
          type: string
          format: url
        website:
          type: string
          format: url
    UserRoleUpdate:
      type: object
      properties:
        newRole:
          type: string
          enum: [user, admin, moderator] # Example roles
          description: The new role to assign to the user.
      required:
        - newRole

    # --- Championship Schemas ---
...
  securitySchemes:
    bearerAuth: # Can be named anything
      type: http
      scheme: bearer
      bearerFormat: JWT # Supabase uses JWT
security:
  - bearerAuth: [] # Applies Bearer token auth globally to all operations by default. Override locally if needed.
paths:
...
  /users/me:
    get:
      summary: Get current user's profile
      tags: [Users]
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Current user's profile data.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserProfile'
        '401':
          description: Unauthorized (user not authenticated).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Profile not found for authenticated user.
    put:
      summary: Update current user's profile
      tags: [Users]
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserUpdate'
      responses:
        '200':
          description: Profile updated successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserProfile'
        '400':
          description: Invalid input data.
        '401':
          description: Unauthorized.
  /users: # Admin endpoint
    get:
      summary: Get all users (Admin only)
      tags: [Users, Admin]
      security:
        - bearerAuth: [] # Requires admin role (checked by backend)
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 10
      responses:
        '200':
          description: A list of users with profiles.
          content:
            application/json:
              schema:
                type: object
                properties:
                  users:
                    type: array
                    items:
                      $ref: '#/components/schemas/UserProfile' # Or a more detailed AdminUserView
                  total:
                    type: integer
                  page:
                    type: integer
                  limit:
                    type: integer
        '401':
          description: Unauthorized.
        '403':
          description: Forbidden (user is not an admin).
  /users/{userId}/role: # Admin endpoint
    put:
      summary: Update user role (Admin only)
      tags: [Users, Admin]
      security:
        - bearerAuth: []
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserRoleUpdate'
      responses:
        '200':
          description: User role updated successfully.
          # Response might include the updated user object or just a success message
        '400':
          description: Invalid input (e.g., invalid role).
        '401':
          description: Unauthorized.
        '403':
          description: Forbidden (user is not an admin).
        '404':
          description: User not found.

  # --- Championship Routes ---
...